more warning fixes

+++ src/fontstash.h
@@ -261,7 +261,7 @@ static void fons__tt_renderGlyphBitmap(FONSttFontImpl *font, unsigned char *outp
 		unsigned int byte_value, num_bits_done, rowstart, bits;
 		for (unsigned int y = 0; y < ftGlyph->bitmap.rows; y++)
 		{
-			for (unsigned int byte_index = 0; byte_index < ftGlyph->bitmap.pitch; byte_index++)
+			for (int byte_index = 0; byte_index < ftGlyph->bitmap.pitch; byte_index++)
 			{
 				byte_value = ftGlyph->bitmap.buffer[y * ftGlyph->bitmap.pitch + byte_index];
 
@@ -957,8 +957,8 @@ static FILE* fons__fopen(const char* filename, const char* mode)
 {
 #ifdef _WIN32
 	int len = 0;
-	int fileLen = strlen(filename);
-	int modeLen = strlen(mode);
+	int fileLen = (int)strlen(filename);
+	int modeLen = (int)strlen(mode);
 	wchar_t wpath[MAX_PATH];
 	wchar_t wmode[MAX_PATH];
 	FILE* f;
@@ -996,7 +996,7 @@ int fonsAddFont(FONScontext* stash, const char* name, const char* path)
 	fseek(fp,0,SEEK_SET);
 	data = (unsigned char*)malloc(dataSize);
 	if (data == NULL) goto error;
-	readed = fread(data, 1, dataSize, fp);
+	readed = (int)fread(data, 1, dataSize, fp);
 	fclose(fp);
 	fp = 0;
 	if (readed != dataSize) goto error;
